requirements.txt
streamlit
python-dotenv
google-cloud-aiplatform
google-auth
google-auth-httplib2
black

-e .
